// frontend-app/src/apiClient/models/TokenHolder.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type TokenHolder = {
    cis2_address: string;
    token_id: string;
    holder_address: string;
    frozen_balance: string;
    un_frozen_balance: string;
    create_time: string;
    update_time: string;
};